fix: Replace 'require' hack with 'silentLoad' function instead
Due to the way R lazy-loads packages, packages are only loaded when one
of their functions are called. The 'oligo' package outputs a lot of
diagnostic information which obscures function output. The 'silentLoad'
function fixes this problem by implementing some of the functionality
in the base R 'library' function and loads the package and its
dependencies silently.